refactor: Decide CLI options for test in JS code
