+++ packages/pyright-scip/package.json
@@ -7,6 +7,8 @@
     "build": "webpack --mode production --progress",
     "clean": "shx rm -rf ./dist ./out README.md LICENSE.txt",
     "prepack": "npm run clean && shx cp ../../README.md . && shx cp ../../LICENSE.txt . && npm run build",
+    "check-snapshots-new": "jest check-snapshots",
+    "update-snapshots-new": "jest update-snapshots",
     "check-snapshots": "npm run update-snapshots -- --check",
     "update-snapshots": "node --enable-source-maps ./index.js snapshot-dir snapshots --environment snapshots/testEnv.json --quiet",
     "test": "jest --forceExit --detectOpenHandles",
