+++ packages/pyright-scip/src/check-snapshots.test.ts
@@ -1,0 +1,3 @@
+import { testMain } from './mainImpl';
+
+testMain('check');
