+++ packages/pyright-scip/src/update-snapshots.test.ts
@@ -1,0 +1,3 @@
+import { testMain } from './mainImpl';
+
+testMain('update');
